requirements.txt
autogen
python-dotenv
requests